+++ lib/cubscout/thread_item.rb
@@ -1,0 +1,4 @@
+module Cubscout
+  class ThreadItem < Object
+  end
+end